hubble: move launchHubble() in the Hubble integration
"attach" launchHubble() to the Hubble cell integration struct. This
complete the removal of the Hubble configuration from the Cilium daemon,
as there is no longer any code under daemon/cmd/hubble.go.

Hubble still depends on the Cilium daemon for flag declaration and
parsing, along with startup.
