+++ daemon/cmd/daemon_main.go
@@ -2002,7 +2002,7 @@ func startDaemon(d *Daemon, restoredEndpoints *endpointRestoreState, cleaner *da
 
 	bootstrapStats.overall.End(true)
 	bootstrapStats.updateMetrics()
-	go launchHubble(d.ctx, d.hubble)
+	go d.hubble.Launch(d.ctx)
 
 	// Start controller to validate daemon config is unchanged
 	cfgGroup := controller.NewGroup("daemon-validate-config")
